fix(docs): update schema references in Swagger file
Update the schema references in the Swagger file to use the correct definitions for the OKRGeneratorRequest and OKRGeneratorResponse200. The previous references were using the incorrect entity names.

- Update the references for OKRGeneratorRequest from #/definitions/entity.OKRGeneratorRequest to #/definitions/okr.OKRGeneratorRequest.
- Update the references for OKRGeneratorResponse200 from #/definitions/entity.OKRGeneratorResponse200 to #/definitions/okr.OKRGeneratorResponse200.

This change ensures that the correct schema definitions are used in the Swagger file.

fix(okr): handle error responses in OKRGeneratorHandler

Handle error responses in the OKRGeneratorHandler function by returning a JSON response with the appropriate status code and error messages.

- If there is an error while parsing the request body, return a JSON response with a status code of 400 (Bad Request) and the error message as the response body.
-