feat(backend): add env-generator command
Add the  command to generate the  file and populate it with environment variables. This command can be used to easily configure the backend application.

The  file is used by the backend application to read the configuration values at runtime. It is generated based on the environment variables specified during the build process.

The  command retrieves the values of the environment variables , , , , and  and writes them to the  file in TOML format.

The  Docker image is now built with additional build arguments , , , , and  so that the values can be passed to the  command during the build process.

Also, update the Docker Compose file to set the default values for the environment variables: , ,