feat(workflows): update Go workflow configuration
Update the Go workflow configuration to specify that it runs on the main branch for both push and pull requests.

In the deploy job, replace the webfactory/ssh-agent action with the appleboy/ssh-action action. This new action provides SSH functionality and allows us to connect to the specified host using the provided private key and username. Additionally, add a step to print the username of the currently authenticated user.