feat(workflow): rename and add new workflow files for build and deploy
This commit renames the existing workflow file  to  and adds a new workflow file .
The  file is renamed to better reflect its purpose of building the project. The new name, , provides a clearer description of what the workflow does.

The new file  is added to the  directory. This file defines a workflow for deploying the project. It runs on the  event and only on the  branch.
The workflow uses the  GitHub Action to connect to a remote server specified by SSH credentials stored as secrets. It then runs the adiatma script on the remote server.

These changes improve the organization of the workflow files and allow for better separation of concerns in the CI/CD pipeline.