chore(deploy): update deploy workflow to build Docker image
Added a new step to the deploy workflow that builds a Docker image with the following build arguments:
- HOST
- PORT
- VERSION
- TOKEN
- GOOGLE_CLIENT_ID
- GOOGLE_CLIENT_SECRET
- GOOGLE_REDIRECT_URL
- GOOGLE_CLIENT_REDIRECT_URL
- SUPABASE_URL
- SUPABASE_KEY

The Docker image is tagged with the GitHub commit SHA and pushed to the Docker Hub repository associated with the specified Docker username. The login to Docker Hub is done using the Docker token and username from secrets.