chore(workflow): update deploy workflow
Update the deploy workflow to use the latest version of the 'docker/login-action' and 'docker/build-push-action' actions. This ensures that the workflow is running with the latest features and fixes provided by these actions.

The changes made in this commit are as follows:

- Updated the 'actions/checkout' step to use version 3.
- Replaced the 'Build Docker Image' step with a new step named 'Build & Push', which uses the 'docker/build-push-action' action.
- Updated the 'Log in to Docker Hub' step to use the 'docker/login-action' action.
- Updated the paths and arguments in the 'Build & Push' step to match the new file structure and environment variables.

These changes were made to improve the efficiency and reliability of the deploy workflow, ensuring that the Docker image is built and pushed correctly according to the specified environment variables.