chore(ci): update Dockerhub login credentials
Update the Dockerhub login credentials in the deploy.yml file. Instead of using the GITHUB_USERNAME and GITHUB_TOKEN secrets, now the deployment process will use the DOCKER_USERNAME and DOCKER_TOKEN secrets for Dockerhub login.

This change is necessary to ensure that the correct credentials are used for pushing the Docker image during the build and push process.