feat(workflow): add pull request trigger for deploying backend
Add a new trigger for the workflow to run when a pull request is made on the main branch. This ensures that the backend is deployed whenever a pull request is created or updated.

In the deploy job, add two new steps to pull the latest Docker image of the backend from the Docker registry and run it in a new container. This ensures that the deployed backend is always using the latest version of the image.

The new steps are as follows:
- Pull the latest Docker image:
- Run the Docker image in a new container: