feat(ci): add deploy-backend workflow
Add a new GitHub Actions workflow file to deploy the backend application. This workflow will be triggered on push events to the main branch.

The workflow includes the following steps:
- Stopping any running Docker containers
- Removing any existing Docker containers
- Removing any existing Docker images
- Pulling the latest version of the backend Docker image from the secret repository
- Running a new instance of the backend Docker container on port 3000

These steps ensure that the backend application is always up to date and running the latest version.

Note: The previous configuration for triggering the workflow on pull requests has been removed as it was not necessary.