feat: add frontend Docker container to docker-compose
Add a new frontend Docker container to the docker-compose.yml file. This container is built from the ./frontend directory using the Dockerfile and sets the BACKEND environment variable to http://localhost:3000. The container is also mapped to port 80 on the host machine.

This change was made to include the frontend in the docker-compose setup and allow it to communicate with the backend server.

Note: This change should be reviewed and tested before merging.