chore(deploy): update deploy.yml
Update the deploy.yml file to fix a formatting issue and use the correct key name for the SSH private keys. The branches should be specified without spaces. This change ensures that the deployment workflow only runs on the main branch. The SSH private key now needs to be provided using the key key instead of password. Additionally, the target directory for the deployment has been updated to . to deploy the docker-compose.yml file to the current directory.