chore(deploy): update deploy workflow, Caddyfile, and Dockerfile
Update deploy workflow deploy.yml by removing extra spaces in 'branches' configuration to match the coding standard.

Update the deploy workflow deploy.yml by changing the name of the 'SCP' step to 'Copy assets to server', for better readability.

Update the deploy workflow deploy.yml by changing the name of the 'SSH' step to 'Re run docker compose', for better readability.

Update the deploy workflow deploy.yml by adding the '--build' flag to the 'docker compose up -d' command, to ensure that the containers are built before starting them.

Update Caddyfile by adding a new domain, api.malascoding.com, and reverse proxying it to the backend service on port 3000.

Update the frontend Dockerfile by splitting it into two stages. The first stage is for building the frontend app and the second stage is for serving the built files with Nginx.

In the first stage, rename