fix(deploy): pull docker images before running
Add a new command in the deploy workflow to pull the latest docker images before running the docker compose up command. This ensures that the latest versions of the images are used in the deployment process.

fix(env-generator): fix typo in environment variable name

Correct a typo in the environment variable name SUPABSE_KEY to SUPABASE_KEY in the env-generator.go file. This ensures that the correct environment variable is used in the application.